Rename fooapp and better calculation
