+++ django_perf_test/settings.py
@@ -37,7 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'fooapp',
+    'ormperf',
 ]
 
 MIDDLEWARE_CLASSES = [
