+++ ormperf/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class FooappConfig(AppConfig):
-    name = 'fooapp'
+    name = 'ormperf'
