+++ tox.ini
@@ -16,7 +16,7 @@ envlist=
 
 [testenv]
 commands=
-    manage.py test fooapp
+    manage.py test ormperf
 
 [testenv:py27-django-1.4]
 basepython=/usr/bin/python2.7
